rename table* templates to table-layout*
