+++ packages/dx-react-grid-bootstrap3/src/templates/table-layout.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
-  TableLayout,
+  TableLayout as TableLayoutCore,
   StaticTableLayout,
 } from '@devexpress/dx-react-grid';
 
@@ -35,14 +35,14 @@ const bodyTemplate = ({ children, ...restProps }) => (
   <tbody {...restProps}>{children}</tbody>
 );
 
-export const Table = ({
+export const TableLayout = ({
   headerRows,
   bodyRows,
   columns,
   cellTemplate,
   rowTemplate,
 }) => (
-  <TableLayout
+  <TableLayoutCore
     layoutComponent={StaticTableLayout}
     headerRows={headerRows}
     rows={bodyRows}
@@ -57,7 +57,7 @@ export const Table = ({
   />
 );
 
-Table.propTypes = {
+TableLayout.propTypes = {
   headerRows: PropTypes.array.isRequired,
   bodyRows: PropTypes.array.isRequired,
   columns: PropTypes.array.isRequired,
